pyproject.toml
[project]
name = "context-critic-ikms"
version = "0.1.0"
description = "Multi-Agent RAG system with Context Critic for intelligent document Q&A"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.124.2",
    "langchain>=1.1.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.3",
    "langchain-openai>=1.1.3",
    "langchain-pinecone>=0.2.13",
    "pinecone>=7.3.0",
    "pypdf>=6.4.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
]
